race track, camera follow
I put in a race track as well as a script for the camera to follow the car. I also imported the car sprite that Jasmine made. We should definitely make our own race track at some point because this one is not very good, but I just put this one in for now to see how it looks with my camera follow.